pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "haystack_extensions"
dynamic = ["version"]
description = 'An set of useful extensions for the Haystack AI library.'
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
keywords = []
authors = [
  { name = "Alexander Gruhl", email = "gruhl.alexander@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "haystack-ai",
  "boto3>=1.28.57",
  "transformers"
]

[project.urls]
Documentation = "https://github.com/Redna/haystack-extensions#readme"
Issues = "https://github.com/Redna/haystack-extensions/issues"
Source = "https://github.com/Redna/haystack-extensions/tree/main/integrations/amazon_bedrock"

[tool.hatch.build.targets.wheel]
packages = ["src/haystack_extensions"]

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
  "haystack-pydoc-tools",
]

[tool.hatch.version]
path = "src/haystack_extensions/__init__.py"

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]
docs = [
  "pydoc-markdown pydoc/config.yml"
]
[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "black>=23.1.0",
  "mypy>=1.0.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive --explicit-package-bases {args:src/ tests}"

style = [
  "ruff {args:.}",
  "black --check --diff {args:.}",
]

fmt = [
  "black {args:.}",
  "ruff --fix {args:.}",
  "style",
]

all = [
  "style",
  "typing",
]

[tool.black]
target-version = ["py38"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py38"
line-length = 120
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
  # Ignore unused params
  "ARG001", "ARG002", "ARG005",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.isort]
known-first-party = ["haystack_extensions"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source = ["haystack_extensions"]
branch = true
parallel = true

[tool.coverage.report]
omit = ["*/tests/*", "*/__init__.py"]
show_missing=true
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[[tool.mypy.overrides]]
module = [
  "botocore.*",
  "transformers.*",
  "boto3.*",
  "haystack.*",
  "haystack_integrations.*",
  "pytest.*",
  "numpy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--strict-markers"
markers = [
  "unit: unit tests",
  "integration: integration tests",
  "embedders: embedders tests",
  "generators: generators tests",
]
log_cli = true